fix(auth): remove redundant remove guest identityId call (#13789)
